requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2

